+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+5.6
+---
+
+ * Deprecated passing more than one attribute to `@IsGranted()`. Use an expression instead
+   (e.g. `@IsGranted("has_role('ROLE_USER') or has_role('ROLE_ADMIN')")`).
+
 5.2
 ---